s/ReadLong/ReadInt/ to match the return type

+++ neo/d3xp/MultiplayerGame.cpp
@@ -4122,7 +4122,7 @@ void idMultiplayerGame::ClientReadStartState( const idBitMsg &msg ) {
 
 	// read the state in preparation for reading snapshot updates
 	gameState = (idMultiplayerGame::gameState_t)msg.ReadByte();
-	matchStartedTime = msg.ReadLong( );
+	matchStartedTime = msg.ReadInt( );
 	startFragLimit = msg.ReadShort( );
 	while ( ( client = msg.ReadShort() ) != MAX_CLIENTS ) {
 		assert( gameLocal.entities[ client ] && gameLocal.entities[ client ]->IsType( idPlayer::Type ) );
@@ -4143,7 +4143,7 @@ idMultiplayerGame::ClientReadWarmupTime
 ================
 */
 void idMultiplayerGame::ClientReadWarmupTime( const idBitMsg &msg ) {
-	warmupEndTime = msg.ReadLong();
+	warmupEndTime = msg.ReadInt();
 }
 
 /*
